fix(db): grava ground speed e track decimais (migracao 003)
O SBS manda ground speed e track com decimal ("157.9", "295.7") e os tipos
Go sao float64, mas a migracao 001 criou aircraft_states.ground_speed,
aircraft_states.track e flights.max_ground_speed como INTEGER. Todo INSERT
com valor fracionario falhava com 'pq: invalid input syntax for type
integer' e o estado era descartado: no receptor, ~87% das MSG,4 (track
quase sempre fracionario), cerca de 26% de todas as mensagens.

A migracao 003_decimal_speed_track converte as tres colunas para DOUBLE
PRECISION (sem perda para as linhas existentes). Ela roda no startup do
tracker, junto das outras; a lista agora vem de migrations.All().

Testes de integracao contra timescale/timescaledb:2.27.1-pg16 (mesma
versao de producao): um grava speed/track/max_ground_speed decimais e le
de volta, o outro aplica 001-002, semeia dados, sobe para a 003 e confere
que os valores e o continuous aggregate aircraft_states_hourly seguem ok.
Sem a 003 ambos falham com o mesmo erro de producao.

+++ internal/db/schema.sql
@@ -7,8 +7,8 @@ CREATE TABLE IF NOT EXISTS aircraft_states (
     hex_ident TEXT NOT NULL,
     callsign TEXT,
     altitude INTEGER,
-    ground_speed INTEGER,
-    track INTEGER,
+    ground_speed DOUBLE PRECISION,
+    track DOUBLE PRECISION,
     latitude DOUBLE PRECISION,
     longitude DOUBLE PRECISION,
     vertical_rate INTEGER,
@@ -37,7 +37,7 @@ CREATE TABLE IF NOT EXISTS flights (
     last_latitude DOUBLE PRECISION,
     last_longitude DOUBLE PRECISION,
     max_altitude INTEGER,
-    max_ground_speed INTEGER
+    max_ground_speed DOUBLE PRECISION
 );
 
 -- Create indexes for flights
